refactor(web): use the shared formatBytes in chat attachments
The two chat-attachment modules each shipped an identical local copy of
formatBytes, while lib/utils/format-bytes.ts is the canonical helper whose
doc comment asks new code to import it.

The copies had no upper clamp (>= 1 TB indexed past the end of the unit
array and rendered "1 undefined") and no NaN/negative guard (NaN rendered
"NaN undefined"). Attachment sizes never reach 1 TB, so the realistic case
is a missing or malformed `size`.

Delete both copies and import the canonical helper at every call site.
Sizes at KB+ now always carry one decimal ("10.0 MB" instead of "10 MB"),
since the canonical helper does not strip a trailing zero; unit and e2e
expectations updated to match. Adds NaN/Infinity and above-TB clamp cases
to the canonical helper's tests.
