+++ apps/web/components/task-create-dialog-selectors.tsx
@@ -8,9 +8,9 @@ import { IconPaperclip } from "@tabler/icons-react";
 import { Combobox } from "./combobox";
 import { scoreBranch } from "@/lib/utils/branch-filter";
 import { BranchRefreshButton } from "./branch-refresh-button";
+import { formatBytes } from "@/lib/utils/format-bytes";
 import {
   processFile,
-  formatBytes,
   MAX_FILES,
   MAX_TOTAL_SIZE,
   type FileAttachment,
